tests: Separate source code and tests
