+++ config/jest.js
@@ -4,5 +4,5 @@ module.exports = {
   resetModules: true,
   clearMocks: true,
 
-  testMatch: [`<rootDir>/src/**/*.test.{js,jsx,ts,tsx}`],
+  testMatch: [`<rootDir>/test/**/*.test.{js,jsx,ts,tsx}`],
 };
