+++ tsconfig.build.json
@@ -5,6 +5,5 @@
     "noEmit": false,
     "outDir": "dist"
   },
-  "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts"]
+  "include": ["src/**/*.ts"]
 }
